Added some changes to code. Sending uart, modified filter settings, small refactor

+++ AutomaticGearShiffter/Include/MagneticSensors.h
@@ -13,7 +13,7 @@ void initializeMagneticSensors(void);
 void handleWheelMagnetInt(void);
 void handleCrankMagnetInt(void);
 void computeVelocityInMetersPerSecond();
-uint8_t getBikeVelocityInKmPerH();
+float getBikeVelocityInKmPerH();
 void computeCadence();
 
 #endif
